requirements.txt
opencv-python>=4.5.0
numpy>=1.19.0
scikit-learn>=0.24.0
xgboost>=1.4.0
lightgbm>=3.3.0
joblib>=1.0.0
PyYAML>=5.4.0
Pillow>=8.0.0 